backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
drf-yasg==1.17.1
gitignore==0.0.8
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==0.16.0
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.19.1
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
static3==0.7.0
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.10
validate-email==1.3
whitenoise==5.2.0
